port authenticate message test
